requirements.txt
requests
python-dotenv
glean-api-client
faker
reportlab